+++ views/err/404.jade
@@ -2,4 +2,11 @@ extend ../templates/framework.jade
 block title
   | 404 Not Found
 block content
-  h1.am-text-center No. There's nothing but air.
+  h1.am-text-center
+    br
+    | Do
+    br
+    | BOOM!
+  h3.am-text-center with
+  p.am-text-center: em=params.url
+  p.am-text-center but he's not here.
